requirements.txt
fastapi
uvicorn
aiohttp
pandas
numpy
scipy
python-dotenv
